requirements.txt
requests
beautifulsoup4
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
huggingface_hub
python-dotenv
serpapi
pandas
